+++ src/api_gateway/api_gateway.ts
