+++ src/api_gateway/api_gateway_config.ts
@@ -1,0 +1,8 @@
+import { apis } from '~/src/api_gateway/config/apis'
+
+export default {
+  apis,
+  modules: [
+    '~/src/api_gateway/modules/oauth2'
+  ]
+}
