+++ src/api_gateway/config/routes.ts
@@ -1,4 +1,4 @@
-import * as c from '~/src/controllers'
+import * as c from '~/src/api_gateway/controllers'
 
 export function routes (route): void {
   route.resources('/users', c.UsersController)
