+++ src/api_gateway/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from '~/src/api_gateway/controllers/users_controller'
