--- src/config.ts
@@ -1,9 +1,0 @@
-import { UserService } from '@iredium/butterfly/lib/services'
-import { routes } from '~/src/config/routes'
-import { databases } from '~/src/config/databases'
-
-export default {
-  userServiceClass: UserService,
-  routes,
-  databases
-}
