--- src/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from '~/src/controllers/users_controller'
