+++ src/index.ts
@@ -1,13 +1,5 @@
-import { ApiGateway } from '~/src/api_gateway'
-import Butterfly from '@iredium/butterfly'
-import config from '~/src/config'
-import apiGatewayConfig from '~/src/api_gateway/config'
+import { Tanhua } from './api_gateway/tanhua'
 
-const butterfly = new Butterfly(config)
+const tanhua = new Tanhua()
 
-let apiGateway = new ApiGateway(butterfly.app, apiGatewayConfig)
-
-apiGateway.init()
-  .then((): void => {
-    butterfly.boot()
-  })
+tanhua.boot()
